implemented tracking of filter status, refactore FilterPanel component
