+++ tslint.json
@@ -12,7 +12,7 @@
   "import-blacklist": [true, "rxjs"],
   "interface-over-type-literal": false,
   "interface-name": false,
-  "max-line-length": [true, 120],
+  "max-line-length": [true, 140],
   "member-access": false,
   "member-ordering": [true, { "order": "fields-first" }],
   "newline-before-return": false,
